Start move important files to core folder
Still need a new reload system that can keep the
reload plugin functional even if reloading fails
