+++ PCBot/plugins/checkers.py
@@ -20,7 +20,7 @@ from miru.internal.types import InteractiveButtonStylesT
 from re import Match, IGNORECASE, search
 from typing import Awaitable, Callable, Optional
 from PCBot.botdata import BotData
-from PCBot.plugins.replyhandler import (
+from PCBot.core.replyhandler import (
   add_game, get_interaction_channel, GuessOutcome, remove_game, TextGuessGame
 )
