+++ PCBot/plugins/hangman.py
@@ -13,7 +13,7 @@ from logging import getLogger, Logger
 from random import randrange
 from string import ascii_lowercase
 from typing import Optional
-from PCBot.plugins.replyhandler import (
+from PCBot.core.replyhandler import (
   GuessOutcome, remove_game, send_text_message, TextGuessGame
 )
