+++ PCBot/plugins/reload.py
@@ -3,7 +3,7 @@
 import crescent
 import hikari
 import logging
-from PCBot.pluginmanager import (
+from PCBot.core.pluginmanager import (
     get_plugin_names, reload_handlers, reload_plugin_manager, reload_plugins
 )
 from typing import Optional
@@ -71,7 +71,7 @@ class ReloadCommand:
         try:
             reload_plugin_manager()
             reload_handlers(plugins)
-            await reload_plugins(plugins, plugin_folder)
+            reload_plugins(plugins, plugin_folder)
             safe_mode = False
         except:
             logger.exception('An error occurred while reloading plugins:')
